pyproject.toml
[tool.poetry]
name = "rokis-corner"
version = "1.0.1"
description = ""
authors = ["Miloš Roknić <roknic.milos.994@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-extensions = "^3.2.3"
django-pdf-view = "^2.1.0"
gunicorn = "^23.0.0"
pillow = "^10.4.0"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
django-smart-fixtures = "^1.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.1"
ruff = "^0.6.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80
